weather_report_etl/requirements.txt
azure-functions
pymongo
requests